Add GitHub Packages auth to .npmrc template and CI workflow #102
