+++ .github/workflows/ci.yml
@@ -1,6 +1,7 @@
 name: CI
 permissions:
   contents: read
+  packages: read
 
 on:
   push:
@@ -30,6 +31,7 @@ jobs:
         with:
           node-version: '24'
           cache: 'pnpm'
+          registry-url: 'https://npm.pkg.github.com'
 
       - name: Setup ESLint cache
         uses: actions/cache@v4
@@ -110,6 +112,7 @@ jobs:
         with:
           node-version: '24'
           cache: 'pnpm'
+          registry-url: 'https://npm.pkg.github.com'
 
       - name: Setup build cache
         uses: actions/cache@v4
