+++ scripts/init-logic-merge.test.ts
@@ -1,0 +1,115 @@
+import { describe, expect, it } from 'vitest'
+import { init_logic } from './init-logic'
+
+describe('merge_json_extends', () => {
+	it('adds extends when key is missing', () => {
+		const result = JSON.parse(init_logic.merge_json_extends('{}', 'my-config')) as {
+			extends: unknown
+		}
+
+		expect(result.extends).toStrictEqual(['my-config'])
+	})
+
+	it('converts string extends to array and prepends entry', () => {
+		const result = JSON.parse(
+			init_logic.merge_json_extends('{"extends":"existing"}', 'my-config'),
+		) as { extends: unknown }
+
+		expect(result.extends).toStrictEqual(['my-config', 'existing'])
+	})
+
+	it('prepends to existing array', () => {
+		const result = JSON.parse(
+			init_logic.merge_json_extends('{"extends":["existing"]}', 'my-config'),
+		) as { extends: unknown }
+
+		expect(result.extends).toStrictEqual(['my-config', 'existing'])
+	})
+
+	it('returns content unchanged when entry already in extends', () => {
+		const content = '{"extends":["my-config"]}'
+		const result = init_logic.merge_json_extends(content, 'my-config')
+
+		expect(result).toBe(content)
+	})
+
+	it('handles tsconfig.json with JSONC line comments', () => {
+		const content = '{\n\t// compiler options\n\t"extends": ["existing"]\n}'
+		const result = JSON.parse(init_logic.merge_json_extends(content, 'my-config')) as {
+			extends: unknown
+		}
+
+		expect(result.extends).toStrictEqual(['my-config', 'existing'])
+	})
+})
+
+describe('merge_json_array_field', () => {
+	it('adds values to an empty array', () => {
+		const result = JSON.parse(
+			init_logic.merge_json_array_field('{"recommendations":[]}', 'recommendations', ['a', 'b']),
+		) as { recommendations: unknown }
+
+		expect(result.recommendations).toStrictEqual(['a', 'b'])
+	})
+
+	it('adds only values not already present', () => {
+		const result = JSON.parse(
+			init_logic.merge_json_array_field('{"recommendations":["a"]}', 'recommendations', ['a', 'b']),
+		) as { recommendations: unknown }
+
+		expect(result.recommendations).toStrictEqual(['a', 'b'])
+	})
+
+	it('returns content unchanged when all values already present', () => {
+		const content = '{"recommendations":["a","b"]}'
+		const result = init_logic.merge_json_array_field(content, 'recommendations', ['a', 'b'])
+
+		expect(result).toBe(content)
+	})
+
+	it('handles extensions.json with JSONC line comments', () => {
+		const content = '{\n\t// extensions\n\t"recommendations": ["a"]\n}'
+		const result = JSON.parse(
+			init_logic.merge_json_array_field(content, 'recommendations', ['b']),
+		) as { recommendations: unknown }
+
+		expect(result.recommendations).toStrictEqual(['a', 'b'])
+	})
+})
+
+describe('merge_package_scripts', () => {
+	const SCRIPT_KEY = 'build'
+	const SCRIPT_VAL = 'tsc --noEmit'
+
+	it('adds missing scripts', () => {
+		const result = JSON.parse(
+			init_logic.merge_package_scripts('{"scripts":{}}', { [SCRIPT_KEY]: SCRIPT_VAL }),
+		) as { scripts: Record<string, string> }
+
+		expect(result.scripts[SCRIPT_KEY]).toBe(SCRIPT_VAL)
+	})
+
+	it('does not overwrite existing scripts', () => {
+		const result = JSON.parse(
+			init_logic.merge_package_scripts(`{"scripts":{"${SCRIPT_KEY}":"existing"}}`, {
+				[SCRIPT_KEY]: SCRIPT_VAL,
+			}),
+		) as { scripts: Record<string, string> }
+
+		expect(result.scripts[SCRIPT_KEY]).toBe('existing')
+	})
+
+	it('returns content unchanged when all scripts present', () => {
+		const content = `{"scripts":{"${SCRIPT_KEY}":"${SCRIPT_VAL}"}}`
+
+		expect(init_logic.merge_package_scripts(content, { [SCRIPT_KEY]: SCRIPT_VAL })).toBe(content)
+	})
+
+	it('creates scripts key when missing from package json', () => {
+		const result = JSON.parse(
+			init_logic.merge_package_scripts('{}', { [SCRIPT_KEY]: SCRIPT_VAL }),
+		) as { scripts: Record<string, string> }
+
+		expect(result.scripts[SCRIPT_KEY]).toBe(SCRIPT_VAL)
+	})
+})
