+++ templates/.npmrc
@@ -1,1 +1,2 @@
 @joshuafolkken:registry=https://npm.pkg.github.com
+//npm.pkg.github.com/:_authToken=${NODE_AUTH_TOKEN}
